requirements.txt
# requirements.txt
pymongo
python-dotenv
faker
